cut class in two
