+++ lib/pages/file_manager_page.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 
 import 'package:flutter/cupertino.dart';
 import 'package:lsa_gloves/datacollection/file_content.dart';
+import 'package:lsa_gloves/navigation/navigation_drawer.dart';
 import 'package:path/path.dart';
 import 'package:flutter/material.dart';
 import 'package:lsa_gloves/datacollection/storage.dart';
@@ -16,6 +17,7 @@ class OldFileManagerPage extends StatelessWidget {
       appBar: AppBar(
         title: Text("Gestion de archivos"),
       ),
+      drawer: NavDrawer(),
       body: SingleChildScrollView(
           child: Column(
             children: <Widget>[
